requirements.txt
Django==5.1.2
djangorestframework
django-phonenumber-field
psycopg2-binary
python-dateutil
pytest
pytest-django
pytest-mockito
factory-boy
django-redis
djangorestframework-simplejwt
django-ratelimit
drf-yasg
python-dotenv
dependency-injector
django-filter